Add stemmer to the tokenizer and change the saving method to pickle whole object.
It turns out that pickling output of the .get_params() does not work (I mean,
the parameters are not the trainned parameters, just configuration).
Thus, pickling whole object.
Can be improved to just pickle the Label encoder and tfidf, leaving the
wrapping class as is, so the interface and parameters are more explicit