Squashed commit of the following: Create the MMDisambiguator class and introduce script for model training and saving.
The model class is the MMDisambiguator.
It uses Logistic Regression classifier. It also holds reference
to the data model so it can transform the predicted classes into labels
'device' and 'animal'. To be consistent with this, it would also be desirable
for the MMDisambiguator class to hold the whole pipeline for feature extraction
from the raw texts, so one could call .predict(sentence) and get back the class
label. Right now it only accepts feature vectors derived from the sentence outside
of the class both for prediction and training.

The script for performing training is src/model/train_and_save_model.py.

The notebooks explore the process of training the classifier, saving it and
loading for prediciotn.

For saving the models, it turns out that the only way is to pickle whole objects
that holds the trained data.
It may be a good idea, to just pickle the insides and leave the classes that mainly
provide interfaces. This way we can build a class and then just supply it with
trained classifiers, instead of pickling the whole class.

Also, added some parameters and reports form training.

commit ee20a93ed34cefe589fe1f57bc4998a42ddb89d8
Author: Damian Kwasny <damiankwasny95@gmail.com>
Date:   Thu Nov 21 14:45:36 2019 +0100

    Update the model class and introduce script for model training and saving.

    The model class is the MMDisambiguator.
    It uses Logistic Regression classifier. It also holds reference
    to the data model so it can transform the predicted classes into labels
    'device' and 'animal'. To be consistent with this, it would also be desirable
    for the MMDisambiguator class to hold the whole pipeline for feature extraction
    from the raw texts, so one could call .predict(sentence) and get back the class
    label. Right now it only accepts feature vectors derived from the sentence outside
    of the class both for prediction and training.

    The script for performing training is src/model/train_and_save_model.py.

    The notebooks explore the process of training the classifier, saving it and
    loading for prediciotn.

    For saving the models, it turns out that the only way is to pickle whole objects
    that holds the trained data.
    It may be a good idea, to just pickle the insides and leave the classes that mainly
    provide interfaces. This way we can build a class and then just supply it with
    trained classifiers, instead of pickling the whole class.

    Also, added some parameters and reports form training.

commit 88861daad854fda34450ec9bea8effd3f5a6858f
Merge: 641fae7 b3497bd
Author: Damian Kwasny <damiankwasny95@gmail.com>
Date:   Wed Nov 20 15:00:56 2019 +0100

    Merge branch 'prototype' into p_model

commit 641fae7289e79201fec376eb22692da17d2750e8
Merge: 046913d 862209f
Author: Damian Kwasny <damiankwasny95@gmail.com>
Date:   Tue Nov 19 13:17:55 2019 +0100

    Merge branch 'prototype' into p_model

commit 046913d8734b3cf0847357e4e43b029e0694930b
Author: Damian Kwasny <damiankwasny95@gmail.com>
Date:   Tue Nov 19 13:13:48 2019 +0100

    Add skeleton for model.

commit 0e5f36b4ac68ae4324a253cc4203839719fb30f5
Author: Damian Kwasny <damiankwasny95@gmail.com>
Date:   Thu Nov 14 17:45:51 2019 +0100

    update docs with model decision info