Refactor TextLabelVectorizer and MMDisambiguator to save collaborators, not whole objects.
Additionally, updated the scripts that save and loads data to comply with the
chamges.
Moreover, updated the documentation of various classes, modules and scripts.

TextLabelVectorizer should be further refactor so it is built from outside, not creating
its own dependants. That way, it could be saved the same way as the MMDisambiguator