Move API creation to builder in api/__init__.py. Add tests for the whole system.
Big update that adds the following:
API is moved top '__init__.py' and now it is possible to create an app by
running the create_app function that is there. Thanks to this change
I could make tests for the API as well.

Tests added for all classes and their functions.

Documentation largely extended.